pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis-research-os"
version = "5.2.0"
description = "JARVIS Research OS - AI-Powered Systematic Literature Review Assistant"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "JARVIS Team", email = "jarvis@kaneko-ai.dev"}
]
keywords = [
    "systematic-review",
    "literature-review",
    "evidence-grading",
    "citation-analysis",
    "PRISMA",
    "active-learning",
    "research-assistant",
    "AI",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "requests>=2.28.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ml = [
    "lightgbm>=4.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
pdf = [
    "pymupdf>=1.22.0",
    "pypdf2>=3.0",
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
    "reportlab>=4.0",
]
llm = [
    "google-generativeai>=0.3.0",
    "llama-cpp-python>=0.2.0",
]
embedding = [
    "sentence-transformers>=2.2.0",
    "rank_bm25>=0.2.2",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "safety>=2.3.0",
    "bandit>=1.7.0",
]
all = [
    "jarvis-research-os[ml,pdf,llm,embedding,web]",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
jarvis = "jarvis_cli:main"
jarvis-screen = "jarvis_core.active_learning.cli:cmd_screen"

[project.urls]
Homepage = "https://github.com/kaneko-ai/jarvis-ml-pipeline"
Documentation = "https://github.com/kaneko-ai/jarvis-ml-pipeline/docs"
Repository = "https://github.com/kaneko-ai/jarvis-ml-pipeline"
Issues = "https://github.com/kaneko-ai/jarvis-ml-pipeline/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["jarvis_core*", "jarvis_tools*", "jarvis_web*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["jarvis_core"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
fail_under = 80

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W"]
ignore = [
    "E501",   # Line too long
    "E402",   # Module level import not at top
    "W293",   # Blank line contains whitespace
    "E712",   # Comparison to True/False
    "F841",   # Unused variable
    "F401",   # Unused import
    "E741",   # Ambiguous variable name
    "W291",   # Trailing whitespace
    "F821",   # Undefined name
    "F402",   # Import shadowed
    "E722",   # Bare except
]

[tool.ruff.isort]
known-first-party = ["jarvis_core", "jarvis_web", "jarvis_tools"]

[tool.ruff.per-file-ignores]
"tests/*" = ["E402", "F841", "E712", "B007"]
"jarvis_core/*" = ["F821", "F402"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
